+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.3
+
+* Add custom systemUiOverlayStyle in Detail Screen.
+* Add enable/disable tap to hide footer and header in Detail Screen. (via tapToHide property)
+
 ## 0.0.2
 
 * Format.